Native methods usage

+++ src/models/Pair.scala
@@ -1,5 +1,5 @@
 package models
 
-class Pair(val point: Point, val distance: Float) extends Ordered[Pair] {
+class Pair(val point: Point, val distance: Double) extends Ordered[Pair] {
     def compare(that: Pair) = this.distance.compareTo(that.distance)
 }
